remove the 'entry' module entirely, moving entry points to submodules of master and outstation
