--- dnp3/src/entry/mod.rs
@@ -1,2 +1,0 @@
-/// entry points for creating and spawning master tasks
-pub mod master;
