+++ dnp3/src/master/mod.rs
@@ -6,6 +6,20 @@ pub mod error;
 pub mod handle;
 /// types related to making requests on an Association
 pub mod request;
+/// entry points for creating and spawning serial-based master tasks
+pub mod serial;
+/// entry points for creating and spawning TCP-based master tasks
+pub mod tcp;
+
+/// state of TCP client connection
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub enum ClientState {
+    Connecting,
+    Connected,
+    WaitAfterFailedConnect(std::time::Duration),
+    WaitAfterDisconnect(std::time::Duration),
+    Shutdown,
+}
 
 pub(crate) mod convert;
 pub(crate) mod extract;
