+++ dnp3/src/master/tcp.rs
@@ -1,8 +1,8 @@
 use crate::app::retry::ExponentialBackOff;
 use crate::config::LinkErrorMode;
-use crate::entry::master::ClientState;
 use crate::master::handle::{Listener, MasterConfig, MasterHandle};
 use crate::master::session::{MasterSession, RunError};
+use crate::master::ClientState;
 use crate::tokio::net::TcpStream;
 use crate::transport::TransportReader;
 use crate::transport::TransportWriter;
