+++ dnp3/src/outstation/tests/harness/harness.rs
@@ -1,6 +1,5 @@
 use crate::config::{AppDecodeLevel, EndpointAddress};
 use crate::link::header::{BroadcastConfirmMode, FrameInfo, FrameType};
-use crate::outstation::OutstationHandle;
 use crate::outstation::config::{Feature, OutstationConfig};
 use crate::outstation::database::EventBufferConfig;
 use crate::outstation::task::OutstationTask;
@@ -8,6 +7,7 @@ use crate::outstation::tests::harness::{
     ApplicationData, Event, EventHandle, MockControlHandler, MockOutstationApplication,
     MockOutstationInformation,
 };
+use crate::outstation::OutstationHandle;
 use crate::tokio::test::*;
 
 use crate::config::LinkErrorMode;
